requirements.txt
Django>=5.0,<6.0
djangorestframework
django-environ
openai
openpyxl
